Corrections and improvement of the early kn filter
stellar mass has been added to the dataframe in `data/mangrove_filtered.csv` for information purposes.
In `filters_early_kn_candidates`: the threshold on angular distances
uses angular diameter distance. The calcul of the absolute magnitude has
been changed in two ways: the magnitude used as apparent magnitude is dc
magnitude and not magpsf anymore. The calcul has also been changed
because distances are in Mpc and not pc.

It appears that the cut on the absolute magnitude is only valid for band
g. However, as alerts mostly come in pair, it seems irrelevant to remove
the cut from band r as it would give a lot of irrelevant alert, when
their g-band associated alert do not pass the cut.