More efficient match
Used `search_around_sky` from astropy to pre-filter galaxies for each alerts.
(run in 20 sec in one night).
Reverted to +-1 for absolute mag error, because as far as I know the error on the expected value is bigger than on the measurements for the most part, but ideally I will check the paper and upgrade that.